+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.3.0 (unreleased)
 ------------------
 
+- #47 Display "Estimated" next to date of birth when estimated from age
 - #46 Added config option to hide age's month and day when older than 1 year
 - #45 Added configuration option for Age introduction
 - #44 Added Sex as a different field from Gender